+++ tests/bootstrap.php
@@ -8,3 +8,8 @@ require __DIR__ . '/../src/CacheListClassLoader.php';
 require __DIR__ . '/../src/FileCacheClassLoader.php';
 
 define('CLASS_BASE', __DIR__ . DIRECTORY_SEPARATOR . 'classes');
+
+function path(array $path, $add = false)
+{
+    return implode(DIRECTORY_SEPARATOR, $path) . ($add ? DIRECTORY_SEPARATOR : '');
+}