+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changelog #
 
+## v4.1.0 (2014-06-19) ##
+
+  * add*Path methods accept a mixed list of paths with namespaces and paths
+    without namespace
+  * Added getBasePaths() and getPrefixPaths() methods to retrieve added paths
+  * Clarified documentation on possibility of using arrays for multiple paths
+
 ## v4.0.1 (2014-05-28) ##
 
   * Code cleanup and documentation fixes